Refactor Dashboard module with BEM rules
