+++ react-app/src/scss/components/_dashboard.scss
@@ -1,0 +1,6 @@
+.dashboard {
+  background-color: #E8EFF3;
+  color: #2E3A42;
+
+  &__content { padding-top: 4rem; }
+}
