--- react-app/src/scss/pages/_dashboard.scss
@@ -1,12 +1,0 @@
-#dashboard-body {
-  background-color: #E8EFF3;
-  color: #2E3A42;
-
-  .dashboard-body-wrapper {
-    padding-top: 4rem;
-  }
-
-  .dashboard-wrapper {
-    padding-top: 1rem;
-  }
-}
